% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgd_fittedTrajectory.R
\name{add_cogs_persubj}
\alias{add_cogs_persubj}
\title{Create a fittedTrajectory object with fits for a single subject}
\usage{
add_cogs_persubj(fit_dat, sex, method = "default",
  xy_pair_name = c("agedays", "htcm"), checkpoints = c(50, 100),
  fun_y_to_raw = who_zscore2htcm, fun_y_to_z = function(x, y, ...)
  return(y))
}
\arguments{
\item{fit_dat}{Subject specific data to be added to the final object.
This must also contain the observed outcome on each subect under a nested dataset "fit", in column named "y".
Similarly, the model predictions must be under the nested dataset "fit", in column named "preds".
Additionally, the holdout and grid predictions can be stored in nested datasets "holdout" and "fitgrid", respectively.}

\item{sex}{The gender (as character "Male" / "Female") for this subject.}

\item{method}{A name of the modeling approach for future model comparison.}

\item{xy_pair_name}{... Same as in hbgd package...}

\item{fun_y_to_raw}{The function to convert the scale of the observed outcome and predictions into the raw scale (height / weight).
If the fits are already on the raw scale then specify this argument as \code{function(x, y, ...) return(y)}.}

\item{fun_y_to_z}{The function to convert the scale of the observed outcome and predictions into the z-scale.
If the fits are already on the z-scale then specify this argument as \code{function(x, y, ...) return(y)}.}
}
\value{
...
}
\description{
Creates objects of class \code{fittedTrajectory} with growth curve predictions, one object for subject ID in the data.
This can be used for further growth curve analysis with hbgd R package.
Requires the the dataset containing the predicted growth curves for all subjects \code{grid_fits_dat}.
}

