% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_features.R
\name{add_fit_plots}
\alias{add_fit_plots}
\title{Add a column to a data frame containing trajectory / velocity plots}
\usage{
add_fit_plots(all_preds, ggplot = FALSE)
}
\arguments{
\item{all_preds}{A data.frame nested by subject and containing fits from \code{\link{predict_all}}}

\item{ggplot}{Set to \code{TRUE} to use ggplot for plotting. Otherwise the default is to use \code{rbokeh}}
}
\description{
Add a column to a data frame containing trajectory / velocity plots
}

