% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{fit_growth}
\alias{fit_growth}
\alias{fit_growth.ModelStack}
\title{Fit Discrete Growth Curve SuperLearner}
\usage{
fit_growth(...)

\method{fit_growth}{ModelStack}(models, method = c("none", "cv", "holdout"),
  data, ID, t_name, x, y, nfolds = NULL, fold_column = NULL,
  hold_column = NULL, hold_random = FALSE, seed = NULL,
  use_new_features = FALSE, refit = TRUE,
  verbose = getOption("gridisl.verbose"), ...)
}
\arguments{
\item{...}{Additional arguments that will be passed on to \code{gridisl::fit_model} function.}

\item{models}{Parameters specifying the model(s) to fit.
This must be a result of calling \code{gridisl::defModel(...) + gridisl::defModel(...)} functions.
See \code{\link{defModel}} for additional information.}

\item{method}{The type of model selection procedure when fitting several models.
Possible options are "none" (no model selection),
"cv" (model selection with V-fold cross-validation), and
"holdout" (model selection based on validation holdout sample).}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points
(in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling.
If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{nfolds}{Number of folds to use in cross-validation.}

\item{fold_column}{The name of the column in the input data that contains the cross-validation
fold indicators (must be an ordered factor).}

\item{hold_column}{The name of the column that contains the holdout observation indicators
(TRUE/FALSE) in the input data.
This holdout column must be defined and added to the input data prior to calling this function.}

\item{hold_random}{Logical, specifying if the holdout observations should be selected at random.
If FALSE then the last observation for each subject is selected as a holdout.}

\item{seed}{Random number seed for selecting a random holdout.}

\item{use_new_features}{Set to \code{TRUE} to use new features (predictors) defined by the growth curve
feature-creator function \code{\link{define_features_drop}}.
Note that the \code{define_features_drop} function is called automatically, but the features
defined inside this function aren't use unless this is set to \code{TRUE}.}

\item{refit}{Set to \code{TRUE} (default) to refit the best estimator using the entire dataset.
When \code{FALSE}, it might be impossible to make predictions from this model fit.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console.
Turn this on by default using \code{options(gridisl.verbose=TRUE)}.}
}
\value{
An R6 object containing the model fit(s).
}
\description{
Define and fit discrete SuperLearner for growth curve modeling.
Model selection (scoring) can be based on either MSE for a single random (or last)
holdout data-point for each subject (method = "holdout") or
V-fold cross-validated MSE which uses entire subjects
(entire growth curves) for model validation (method = "cv").
}
