% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{SLfit_wrapper}
\alias{SLfit_wrapper}
\title{Wrapper for growthcurve SL function calls.}
\usage{
SLfit_wrapper(method, fold_column = NULL, hold_column = NULL, data, ID,
  t_name, y)
}
\arguments{
\item{method}{The type of model selection procedure when fitting several models.
Possible options are "none" (no model selection),
"cv" (model selection with V-fold cross-validation), and
"holdout" (model selection based on validation holdout sample).}

\item{fold_column}{The name of the column in the input data that contains the cross-validation
fold indicators (must be an ordered factor).}

\item{hold_column}{The name of the column that contains the holdout observation indicators
(TRUE/FALSE) in the input data.
This holdout column must be defined and added to the input data prior to calling this function.}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points
(in days, weeks, months, etc).}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{nfolds}{Number of folds to use in cross-validation.}

\item{hold_random}{Logical, specifying if the holdout observations should be selected at random.
If FALSE then the last observation for each subject is selected as a holdout.}
}
\description{
Predefine a wrapper for model fitting.
This function returns a function that takes in two arguments: \code{models} and \code{x}.
See the arguments \code{models} and \code{x} in \code{fit_growth} function for additional details.
}
