% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_all}
\alias{predict_all}
\title{Obtain all possible predictions with one call}
\usage{
predict_all(modelfit, newdata, add_holdout = TRUE, add_grid = TRUE, tgrid,
  grid_size = 150, add_checkpoint = TRUE, checkpoint = as.integer(c(1,
  hbgd::months2days(1:24))), add_MSE = TRUE,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_growth}}.}

\item{newdata}{Subject-specific data for which predictions should be obtained.}

\item{add_holdout}{Optional flag, set to \code{TRUE} to add holdout (out-of-sample) predictions
used for scoring the current best model fit.}

\item{add_grid}{Optional flag, set to \code{TRUE} to add a grid of equally spaced
predictions over some range of the time variable for the current best model fit.}

\item{tgrid}{Specify the grid of time-points directly.
If missing a subject-specific grid is defined based on the subject's follow-up range.}

\item{grid_size}{How many time-points should be used in equally spaced grid?
This argument is only used when \code{tgrid} is missing.}

\item{add_checkpoint}{Set to \code{TRUE} (default) to obtain predictions for pre-specified x checkpoint
(essentially the same thing as grid, but with potentially different spacings).}

\item{checkpoint}{The grid of checkpoint for which to obtain predictions (daily resolution).}

\item{add_MSE}{Add subject-specific MSE for growth curve prediction (NOT IMPLEMENTED).}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to
the console.}
}
\value{
A data.frame with one row per subject.
The column named "fit" contains nested subject-specific observed data and predictions.
Each entry (cell) of "fit" is a list of 4 subject-specific data.frames,
named "xy", "fit", "fitgrid" and "holdout",
containing the training time-outcome values, the model predictions for training data,
the grid of equally spaced model
prediction and the holdout / out-of-sample predictions, respectively.
}
\description{
This function creates a \code{data.frame} with one row per subject and a column named "fits".
The latter column contains the 3 possible subject-level predictions:  fit - training data model predictions,
fitgrid - grid-based model predictions, holdout - predictions for holdouts or out-of-sample predictions.
Some of these predictions might be empty, e.g. when \code{add_holdout = FALSE}, no holdout predictions are made.
}
