% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_all}
\alias{predict_all}
\title{Obtain all possible predictions with one call}
\usage{
predict_all(modelfit, newdata, add_holdout = FALSE, add_grid = FALSE,
  tmin = NULL, tmax = NULL, incr = 5,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_growth}}.}

\item{newdata}{Subject-specific data for which predictions should be obtained.}

\item{add_holdout}{Optional flag, set to \code{TRUE} to add holdout (out-of-sample) predictions
used for scoring the current best model fit.}

\item{add_grid}{Optional flag, set to \code{TRUE} to add a grid of equally spaced
predictions over some range of the time variable for the current best model fit.}

\item{tmin}{Min t value for predicting the entire growth curve.
By default the lowest observed value in \code{newdata} is used.}

\item{tmax}{Max t value for predicting the entire growth curve.
By default the highest observed value in \code{newdata} is used.}

\item{incr}{Increment time variable value for predicting the entire growth curve.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to
the console.}
}
\value{
A data.frame with one row per subject.
The column named "fit" contains nested subject-specific observed data and predictions.
Each entry (cell) of "fit" is a list of 4 subject-specific data.frames,
named "xy", "fit", "fitgrid" and "holdout",
containing the training time-outcome values, the model predictions for training data,
the grid of equally spaced model
prediction and the holdout / out-of-sample predictions, respectively.
}
\description{
This function creates a \code{data.frame} with one row per subject and a column named "fits".
The latter column contains the 3 possible subject-level predictions:  fit - training data model predictions,
fitgrid - grid-based model predictions, holdout - predictions for holdouts or out-of-sample predictions.
Some of these predictions might be empty, e.g. when \code{add_holdout = FALSE}, no holdout predictions are made.
}

